Remove objectMode from FileReadStream
I believe we needed objectMode in the past because the readable stream
could emit data that was ArrayBuffer or Uint8Array, etc.

Now that we always emit data as Buffers, objectMode is not necessary.

This PR doesn't change the FileWriteStream case since users might want
to push various types (besides strings and Buffers) to the stream.