Streaming: Don't load the entire file into memory
Before this change, FileReadStream would read the full file into memory
with a single readAsArrayBuffer() call.

This commit calls readAsArrayBuffer() with a slice of the file blob for
each call to _read(). This means the blob is never fully loaded into
memory.

The same FileReader object is reused, just with different slices of the
file blob passed into readAsArrayBuffer().

Also, this eliminates the event named 'readable' which was causing bugs
when I piped this stream into another stream library.